Tolerate extensions with no natives in them
This modifies the extension scanning script so that it will still
work even if there are no natives.  This is useful if the extension
has side effects (registering objects or codecs) but doesn't have
any natives in it.

It also hooks the parser to make the header information available that
is scanned out of the C comments at the top of the file.