Replace NULL slots in dispatch tables with failing handlers
There are several tables which are indexed by REB_XXX type and have
some kind of handler.  The dispatch for some of these tables would
look in a slot and check for NULL, and if the slot was NULL then
the caller would trigger a failure.

As a slight improvement on this, rather than put a NULL in the slot
this instead registers a function that actually raises the error.
That allows dispatch to avoid the NULL check.

(This feature was initially added in order to avoid making type
dispatch more expensive just because some types didn't have handlers,
even though they all do today--the idea was to make the type dispatch
table mutable, start with a failing handler for REB_GOB and REB_STRUCT
and let an extension load in the replacement handler.  Being able to
put GOB! and STRUCT! into an extension is currently a little far out
however, since Cmp_Value and Mold_Value do not have an extension hook.
Those need to be thought out more before the change is made, which
may be influenced by the user-defined type design.)

Additionally, this renames the files containing tables to ".inc"
files instead of ".h" files, as they can be included only once in
the entire build.