Codecs as FUNCTION!s, can be C natives or custom usermode code
This eliminates the cryptic and error prone REBCDI*-based interface
for codecs, switching to instead consider a codec to consist of
three separate functions (with appropriate type signatures)

    identify? => takes a BINARY!, returns a LOGIC!
    decode => takes a BINARY!, returns ANY-VALUE!
    encode => takes ANY-VALUE!, returns BINARY!

As long as the function signatures are correct, it does not matter
whether the behavior is implemented in Rebol code or C code.

To register a codec, say:

    sys/register-codec*
        'My-Codec-Name
        [%my %my2 %my3] ;-- can be a single file extension
        :my-identify? ;-- can be blank
        :my-decode ;-- can be blank
        :my-encode ;-- can be blank

The codecs are dispatched by LOAD, which is also usermode code in
the SYS context.  If one is unhappy with the way it's working, then
it's now all tweakable as Rebol.

While this was being done, it also incidentally fixed the loading of
animated GIFs (they are returned as a BLOCK! of images), as well as
correcting a memory leak when JPEG files fail to load or an
identification request for a JPEG returns false.