b-boot.c => tmp-boot-block.c, simplify Rebol part of Init_Core()
Looking at the "boot block", it had an empty section for PROTOCOLS.
This is apparently some unused thing that does not correspond to
"schemes".  So it was removed.

At the same time, this moves the call to INIT-SCHEMES into the host.
Ideally it would be finer granularity than "load all the schemes"
and their loader would not be hardcoded into the SYS context.
Instead, each port scheme would be loadable independently as an
extension...but this is one step closer to that.

There's been an evolving effort to make sure that any temporary files
in the build are indicated so with the prefix %TMP.  Since %b-boot.c
was an exception, this changes that to the more obvious naming
of %tmp-boot-block.c - which shares a name with its Rebol source
equivalent %tmp-boot-block.r.

Also brings some comments up to date.