Use LODEPNG-based encoder for PNG, not %u-png.c version
Comments in the code indicate that Saphirion had trouble with some
PNG encodings crashing.  Rather than figure out what was wrong with
the R3-Alpha %u-png.c, the LodePNG encoder was linked in.

The FUNCTION!-based codec change removed a hack in ENCODE which
switched to use the LodePNG-based TO-PNG function instead of the PNG
codec.  This restores the prior use of LodePNG for encoding by just
passing a different encoder in when registering the codec than the
one in %u-png.c

A decision should be made about whether to use LodePNG or update the
%u-png.c sources, or some other choice.  (It will matter less when the
code is an extension--possibly allowing use of built-in decoding on
whatever platform/runtime the interpreter is running on.)