Tighten error reporting and enforce correct lengths for Scan_XXX
The scanner in R3-Alpha was very twisty and "hand-made", and did not
have a fully consistent convention for how it worked.  Most of the
Scan_XXX routines took in a byte pointer and returned a byte pointer,
where the idea of the returned byte pointer was at the end of the
scanned material.  This made parts of the code that weren't the
scanner able to reuse the routine--in theory.

But in practice, there was not much enforcement that the returned
pointer was actually at the position it claimed to be.  Sometimes the
pointer was just compared against NULL, and sometimes even that check
was omitted.  The way errors were handed back to the main scanner
varied, even to checking for an uninitialized VAL_TYPE() in the
returned REBVAL.  :-(

This adds a bit of vetting to the process.  It also changes the
convention of using REBINTs to hold the tokens (with negative token
numbers to indicate an error) to putting the error in the scan state
and making the token an actual enumerated type, which helps with
seeing what's going on in the debugger.