Crude draft of ADD and SUBTRACT on BINARY!
Arithmetic operations should be allowed on BINARY!, because it's too
limiting to not allow `#{4B} + 1` => `#{4C}`.

Allowing the operations requires a default semantic of binaries as
unsigned arithmetic, since one does not want `#{FF} + 1` to be #{FE}.

It uses a big endian interpretation, so `#{00FF} + 1` is #{0100}

Since Rebol is a language with mutable semantics by default, `add x y`
will mutate x by default (if X is not an immediate type).  Today `+`
is a synonym for ADD, but after some technical changes are made to
permit TIGHTEN of ADAPTed functions, it will be an enfixing of ADD-OF,
which will copy the first argument before adding.

To try and maximize usefulness, the semantic chosen is that any
arithmetic that would go beyond the bounds of the length is considered
an overflow.  Hence the size of the binary will not change.  This means
that `#{0100} - 1` is #{00FF}, not #{FF}.

!!! The code is extremely slow and crude--using an odometer-style
loop to do the math.  What's actually required is effectively "bigint"
math, and it might be that it would share code with whatever big
integer implementation was used; e.g. integers which exceeded the size
of the platform REBI64 would use BINARY! under the hood.