Get rid of DELECT
The DELECT native was a not-well-understood "dialecting middleware"
that was included in R3-Alpha.  It was kept in a compiling state in
Ren-C since it seemed like it might be used by R3/View.  However,
despite it compiling there were no tests and nothing to ensure it would
stay in working condition.

When it was discovered that R3/View did not use it (rather, alternative
UI like R3-Gui used it), that combined with the fact that it is almost
certainly not working makes it best to just delete it.