Remove poor optimization from Do_Next_In_Frame_May_Throw
There were some interesting optimization ideas which were tried when
doing a single unit of evaluation into an output slot.  Yet they were
not pulling their weight--seemingly even slowing the code down
slightly in real-world tests.

To simplify matters as the evaluator is being redesigned for complete
handling of left-hand-arg quoting and tightness vs. lookback deferment,
the added complexity is removed for now--turning it into more straight
line code with fewer branches (which is probably why it comes out as
good or better than the optimization).