Legitimize the "tight" convention as new parameter class
After significant philosophical debate, it was determined that there
is likely no getting around the need for having a parameter class
which is evaluative like normal parameters, but stops evaluation at
the point of seeing a function which acquires its first argument from
the left hand side.

This convention was only seen in R3-Alpha in OP!.  However, Ren-C has
embraced functions which get their first argument from the left hand
side that do not wish to associate that with a difference in how they
gather their right hand arguments.  It also has a difference in
wanting to have operations which "defer" their lookback arguments,
e.g. the precedence manipulators:

    x: 1
    y: 2
    z: (x + y <| print "Stuff" print "Afterward")

The <| wants `x + y` to fully evaluate, then run the successive
prints, then have the entire expression evaluate to 3.  This is
now considered the "normal" behavior, while the "tight" version of
this behavior on the left hand argument would be evaluative of y, but
aggressively run the enfixed function and then pass the whole result
as the right hand argument of `+`.

The new parameter convention joins "normal" (WORD!), "hard quote"
(GET-WORD!), "soft quote" (LIT-WORD!), and "local" (SET-WORD!).  The
preliminary choice for denoting a "tight" parameter is with ISSUE!.

This commit also adapts the behavior with variadics.  This means that
one is able to get traditional OP!-like behavior out of a variadic,
which will perceive the rightward-looking view of an enfixed operator
to be an `<end>` (as if it had hit an expression barrier, end of block
or end of group)

    sum-normal: function [x [integer! <...>]] [
        sum: 0 | while [not tail? x] [sum: sum + take x] | sum
    ]

    sum-tight: function [#x [integer! <...>]] [
        sum: 0 | while [not tail? x] [sum: sum + take x] | sum
    ]

    >> sum-normal 1 2 3 * 10
    == 33

    >> sum-tight 1 2 3 * 10
    == 60