Simpler/Better "Facade" (e.g. can use TIGHTEN on any function)
This is a change designed to permit flexibility in rewriting the
parameter conventions of function compositions.  The need was first
imagined with REDESCRIBE (to create adaptations and specializations
with more restrictive type conventions than the functions they were
composing, such as a specialization of APPEND that would only work
on STRING! series).  Yet the technical difficulty of actually
doing such a transformation wasn't seen until TIGHTEN tried to change
the parameter conventions of an arbitrary composition.

Now it is possible to TIGHTEN any function, including chains and
specializations and adaptations.  It can do this without needing to
copy or rebind the function body, and the resulting new function will
not add any runtime overhead beyond what the original cost.  It paves
the way for doing more generic type signature aliasing in the future.

How it works:

When the original experiments like SPECIALIZE and ADAPT were being
written, the concept was that an existing piece of implementation
for a function could be reused by another function, which would get
a chance to run before that code.  But in order for the actual code
of the function to work, the frame pushed had to have the right
number of slots and arguments to match the expectations of the code
that ultimately ran.

(e.g. if you specialize APPEND to make it append a certain value, say
append-x, then that may have only one argument instead of append's 2.
However, the frame that's pushed must be suitable for APPEND and have
room for both arguments, even though the specialization's paramlist
has fewer)

Initially as a performance optimization, rather than have each function
call walk down the chain of compositions to find the "underlying"
function, it would cache the underlying function's pointer in the
Reb_Series of the paramlist of the composition.  This optimization
faced a problem though--since not only was the underlying paramlist
needed but also the most underlying values for specialization.  That's
because while the outermost function was fulfilling the innermost
parameter list, it needed some way to realize which of those inner
parameters it was supposed to be skipping.

That meant looking up the underlying function wasn't always one step
to the lowest level function.  It was sometimes one step to the highest
level specialization first.

Adding to the complexity was an increased desire to allow composed
functions to tweak the parameter types accepted by the composition.
If the parameter specification being used was the one of the underlying
function, it would not be possible to change those types without
affecting the underlying function as well...this meant the chain of
function references would have to be laboriously copied.

This works by storing the specialization exemplar in a separate unused
slot in the function body's series, and then instead of storing an
"underlying function" it stores a "facade".  This facade is an array
which may or may not be the paramlist of the underlying function...
either way it will have a REB_FUNCTION in its [0] slot corresponding
to whatever the underlying function is.  However it may also be an
equally sized array which is not a valid function paramlist, that has
different parameter specifications but the underlying function in the
zero slot.