Support extensions as dynamically linked libraries
Extensions (collections of modules) can now be built as dynamic
linked libraries (.dll, .so or .dyn etc), and can be loaded by
LOAD-EXTENSION, and unloaded by UNLOAD-EXTENSION.

When the extension is loaded, an object will be inserted into
system/extensions. This object has a reference to the open library! that
backs up the extension and other information. All of the modules will be
inserted into system/modules by LOAD-EXT-MODULE.

When the extension is unloaded, all of the modules will be removed, and
the native functions in the modules will be modified so that calling to
these functions will result in an error (which shouldn't be called in
the first place, but if they are, no crash at least).

An extension can also be built into the executable, in which case, it
will be automatically loaded as the interpreter starts up.

Because an extension typically calls the functions provided by the core,
and a standalone executable is desired, all of the core
functions/variables are exported in the executable (Usually,
executables, unlike dynamic libraries, are the final target, and do not
export symbols).

An extension can have a startup script that will run after the
extension is successfully loaded. The script could be saved in the
extension in plain text or compressed form.

Currently make-make doesn't support building the extension as external
libraries, they have to be builtin or not built at all.