fail() clears partial outputs, fix Ctrl-C Heisenbug in console
When the evaluator is called, it is given the memory address of where
the result value is to be written.  This address is known to the frame,
and the list of frames is known to the garbage collector--which allows
this arbitrary location in memory to keep any series it refers to
during the evaluation alive.

One issue with this that had not been addressed was what would become
of the bit pattern at that location in the event of an evaluation
that did not finish.  This could happen either due to a fail() partway
through, or something like a Ctrl-C.

This formalizes the decision such that an END marker is written into
every output slot when unwinding the frame stack, up to the level
that traps it.  A reason for choosing END instead of trash is because
the slot may be a function local which is guarded by GC, and the
caller shouldn't have to be burdened by the potential introduction of
an unsafe value in such slots by evaluation.  The use of END instead
of unreadable blank is since these slots are not supposed to be in
the middle of blocks or in the data stack, so writing END helps add
an extra check to alert callers to if they have done this.