Update copyrights, remove individual credits
Per policy outlined in CREDITS.md, those wishing individual credit for
things in the project should do so in that file.  Otherwise it becomes
a problem after the files are edited by many parties, regarding the
questions of credit (or much more likely, blame) for what's in the file.

Also bumps copyright dates to 2017 where applicable.