Move break handling out of Open_StdIO into host, fix Windows Ctrl-C
When earl factored the console code two years ago, the handling of
Ctrl-C on Windows was made conditional on the detected file type of
stdin.  In addition to the detection failing for some reason, this
may or may not be what is desired by the host to drive the registration
of the Ctrl-C handler.  There are also other forms of breaking which
might need to be handled differently.

Rather than very opaquely hide the Ctrl-C handling, this makes it much
more obvious by putting it as part of the host's main(), with a
conditional #ifdef.

There was some code that injected a textual noop into the input stream
as part of the Ctrl-C handling.  It only did this in the non-Windows
build, and it looked like a poor solution to a problem that predated
the usermode console implementation.  This deletes that as well.

Also fixes C++ build issue from last commit.