Add SLEEP function not dependent on WAIT's port mechanics
Currently, in an AWAKE handler it is not possible to WAIT.  @GrahamChiu
needs to do this, so this just hooks to the system sleep facility
directly through SLEEP.

This is a temporary workaround, pending a better analysis.

Also contains some changes allowing the inclusion of Windows.h after
sys-core.h, by undefining some conflicting #defines.