Get rid of Set_Root_Series
The Set_Root_Series function obscured the type of the argument series,
and wasn't doing anything other than initializing either a block or
a string...when the caller already knew which it was.  It had an
assert that string series were either 1 or 2 bytes wide, but that
check is more appropriate to do in series initialization for all cases.

Also uses the data stack to create the right number of values for the
typeset array, rather than guess at "40".