Kill SET-SCHEME/Register_Scheme() and fixed-size table, use natives
R3-Alpha had a very convoluted way of hybridizing some native C code
with user code for implementing port schemes.  The idea was that
a table would be built using Register_Scheme(), and then this table
would provide "functions" (which weren't really functions) to be
the "port actor" for that scheme.

The reason they weren't really FUNCTION!s is because their C native
function was actually a REBPAF, or "port actor function".  This was
parameterized with the port and the action, and unlike a FUNCTION! the
frame it would expect would vary based on which action it was.

This replaces the non-obvious code with a more obvious solution.  A
native is made to provide the usermode port registration code with a
HANDLE! representing the REBPAF.  This way, scheme registration can be
written as Rebol code, which only needs to ask for the special port
actor.  The existing mechanism for registering a variable number of
natives replaces the need to come up with another way of making a
table of the native REBPAF handles which are to be exposed.

This paves the way for the ports themselves being migrated to be
extensions.