Get rid of C shim for Register_Codec, fix BMP and JPEG registering
This helps streamline %b-init.c by removing the C convenience routine
for registering codecs, which was just an interface to the Rebol
helper routine.  Code wishing to register codecs should just be able
to make the Rebol call (already being done with BMP/JPEG/PNG/GIF
codec extension).

It also fixes the missing provision of ENCODE-BMP to the codec, and
gives the JPEG codec the "JPEG" word name instead of "JPG" (which has
been the historical name referenced in the tests).