Rename Rebol pointer "guessing" to "detection" to denote reliability
The uncertainty in the pointer guessing has been removed, and it is
slated to be depended upon as giving the right answer for what kind of
pointer in the system a (valid) pointer is.  Given that, the name
"guess" is a bit unsettling, as well as having the routine appear to
be in a module related to panics/crashes.

This moves the routine and names it "Detect_Rebol_Pointer".