Capture "specifier derivation" pattern as an inline function
When descending into an array, the values cannot be Derelativize()'d
without a specifier.  And the specifier can change at each level of
descent--though it may not.  So there was a simple derivation rule...
if the child value was specific, switch the specifier to its specifier,
else use the one you already had.

This pattern will need to be adapted for virtual binding, because the
rule for derivation is not as simple...a composite specifier will need
to be generated in some cases (if trying to virtually bind something
that is already bound relatively or virtually).

Even before that happens, it's useful to factor out the points where
a derivation is performed, just to avoid repeating a code pattern.
This does that with Derive_Specifier().