Kill RL_APIs using REBSER/REBARR/REBCTX or REBVAL by non-pointer
The goal of the external API is to decouple the internal concerns
(guarding series from GC protection, for instance) from the needs of
a C codebase which wants to operate at about the same level as writing
Rebol code directly in the evaluator.

To accomplish this, clients of the external API cannot store REBVALs
in memory that the interpreter doesn't know about.  Instead, it has to
use an abstract handle.  (For convenience, Ren-C is making this
abstract handle a void* that can be used internally to the system as
a REBVAL, but it actually is part of a "pairing" series node, containing
an additional "meta" value indicating the lifetime--such as a refcount
or a FRAME! to aattach the lifetime to.)

The old interface of the RL_XXX API was kept on life support for some
time, to support the cryptography and PNG code, that had been written
prior to open sourcing.  These have been converted to extensions, and
anything wishing to have "native level" access to Rebol should be
using the routines they want directly.  Anything wishing for a higher
level of abstraction should not have access to REBSER, REBARR, REBCTX.
So this deletes those from %reb-defs.h when REB_DEF is not defined.

Also, the definition of REBVAL is turned into a void, so that it is
not possible for clients to allocate cells for REBVALs to be stored in.

The definitions of REBGOB and REBEVT still have some lingering struct
fields that are technical REBSERs.  The hope is that both of these
will be able to be managed efficiently as ordinary REBVALs or user
defined types.  For now, they conditionally define their few REBSER
fields as void*, making it equivalent to how it was done before without
a more broad exposure of REBSER in the API.

The end result is jostling a couple of host kit APIs which need better
solutions, which are much more practical to conceive of now.

Additionally, this removes the concept that Rebol exports its own clone
of printf for some reason as RL_Print.