Fix extension passing error code by REBVAL not REBVAL*
C does not have type checking for variadic arguments, but errors for
now only support fully specified REBVAL*.  The error message for
extensions was passing an REBVAL by value instead of by pointer.

Factor code so there's only one place issuing the error message, only
one place doing the closing of lib, and make it clearer that a function
is being called (it looks a little like ret is a cast result to a
QUIT_FUNC if one is not reading carefully).

Also update formatting.