Don't use full REBVAL cell to store global END marker
This adds some C++ destructor behavior to check that a REBVAL cell
is not garbage when it goes out of scope.  A trick previously used
which wrote a non-cell signature into a REBVAL (to make an "unwritable"
END for END_CELL) broke this.  Changing that non-cell signature to
just be stored in a REBNOD rather than a full cell solves the issue.