Disable direct REBVAL assignment, require use of Move_Value()
This uses the C++ build to catch instances of direct assignments of
a REBVAL to another, e.g.

    void Some_Func(REBVAL *dest, const REBVAL *src) {
       *dest = *src;
    }

The reason is to try and have some choice in what is done as part of
the assignment, based on the nature of the destination.  The hope is
to be able to sense when cells are stack-based, and if they are, to
be less aggressive about generating "reified" entities that will need
garbage collection when writing to a stack cell that will be going
away before the reification would make a difference.

The name `Move_Value` is chosen for the operation, but in the "memmove"
sense and not really the `C++ move constructor` sense.  It is a form
of smart copying, but calling it `Copy_Value` could lead to a
misunderstanding...e.g. if the value is a REB_BLOCK then thinking that
the array underyling the block would be copied.  Using a stranger name
like "Move" helps avoid that interpretation, but also cues to
realizing that the value is going from one cell to another and may
have to be recontextualized.

Though it is a little bit "wordier", it actually doesn't look bad