Add DECLARE_LOCAL hook point to replace raw REBVAL declarations
Going along with having a generic hook for noticing the assignment
of one specified REBVAL to another, this is a hook for the point of
declaration of REBVALs in the C stack.

A form of initialization code, similar to what was done in the C++
debug build, is now applied to all builds.  This means that if one
uses a raw REBVAL declaration instead of DECLARE_LOCAL, the slot
will generate an alert as being uninitialized for writing.

Having the hook available is generically useful for debugging, since
code can be injected at the creation of every stack value.  It also
actually makes the code a little nicer, by removing the need to take
the address of the stack values, making behavior more interchangeable
with other value pointers living in the heap.

This code simulates the pattern which is believed to be useful for
tracking the lifetime of stack value slots.  Consequently, each time
DECLARE_LOCAL is run it actually makes two REBVALs.  This mimics the
layout of the REBSER-based "pairings" which are intended to be used for
GC-able API handles, although these are not GC'd.  There is a
pointer definition as an alias to the value's location, but it is
const and not changeable.  So long as the address of the address is
not taken, the optimizer should throw it out.

Because there is initialization code, performance-sensitive routines
should put DECLARE_LOCAL outside of loops.

VALUE_FLAG_STACK is added, but at this point in time it does not do
anything.