Don't require Derelativize or Move_Value destinations to be specific
It has been up for debate whether "output" slots should be relative or
specific, based on the documentation value of the fact that an
overwritten output slot may be typed as a RELVAL but it will be
guaranteed REBVAL when the assignment is over.

However, with Move_Value() being an analogue to plain assignment,
which permitted assignment of specific values to relative...it seems
that convenience should win.  This removes the need to promote
relative values through SINK() to "blank" specific values, and instead
adopts a convention of returning a specific value pointer which may
be used or not.

That seems like better documentation, and since these are inline
functions, that will be optimized out if not used.