Make cell preparation not just debug, simplify value macros
The principal part of this change is to make it so that resetting a
value's header, or moving it, honors the status of the target cell's
VALUE_FLAG_STACK and doesn't overwrite it.

To make this change easier and the code clearer, this pipes more of the
initialization through common calls to VAL_RESET_HEADER, by adding a
VAL_RESET_HEADER_EXTRA variant that passes in more flags.  This lets
header-setting routines be macros and not have to repeat the `__FILE__`
and `__LINE__` capturing in things like SET_BLANK or SET_TRUE.

Additionally, this makes a simplifying change to treat the absence of
NODE_FLAG_VALID as the "trash" bit, rather than using a void value with
a flag set.  This allows the unreadable blank to actually be a blank in
the debug build, which makes the code more clear.