Add basic outline of Move_Value, share header-reset/flag-set code
This corrects a debug build problem which came from not filtering out
the type information in flags like WORD_FLAG_BOUND before OR'ing it
in with the rest of the data in VAL_RESET_HEADER_EXTRA.  In addition
to correcting the problem, it also makes the debug build more robust
by getting the checking done by the shared macro.

It also includes a basic outline of how Move_Value would work if some
updates were done.  This shows how cases could be shortcut and not
require determining the relative stack values.

Includes a try to make the C++ build statically link libgcc as well
as libstdc++, plus adding an HTTPS read request to the "smoke test"