Patch assert in FFI code, allow linkernames not at head
The FFI hadn't been tested in a while, and needed to have when it
initialized the "meta" link of the paramlist moved to before the use
of that paramlist in Make_Function, due to a changed invariant.

While fixing that, this also makes it so that when ARG(name) is passed
in as the name to look for in the DLL or shared library, that it will
honor when the beginning of that string is not at the head of the
series.

With these modifications %gtk.r runs again.