Get rid of optimization using D_CELL for single argument functions
In order to avoid pushing a chunk when a single-argument function was
called, there were tests for 0 and 1 argument functions on both the
push and pop side, which would special-case the reuse of the frame's
temporary cell for the arguments.

This may or may not be an advantage, given that pushes and pops to the
chunk stack are intended to be very fast, and it created more
conditional code.  It also took away use of that temporary from single
argument functions like ANY, which might have a use for them.

Given the speculative idea that stack values will be searchable to find
the REBFRM* to which their lifetime is tied, having a different
argument pushing convention for this is more trouble than it's worth,
when it probably wasn't all that useful to begin with.  So this takes
it out.