Simplify implementation of DO in PARSE, non-END subparse targets
While trying to pin down some of the workings of how layers of PARSE
were returning their results in R3-Alpha, if a function returned a
result (like an index, END_FLAG, or THROWN_FLAG) then it was kept
from being inconsistent with the value cell that holds the actual
output, by ensuring the output was only written with a thrown cell,
or at the edges of the native operation.  This changes that assert to
be more purposeful, by actually asserting it in the routines and not
asking callers to pass in an END marker as the initial target.

Additionally, this simplifies the DO operation in PARSE, so that it
reuses more of the common parsing.  DO is a somewhat problematic
operation in general, and it reveals some of the weaknesses of Rebol's
ad-hoc PARSE implementation methodology.  This doesn't fix any of that
in a meaningful way, it just reduces the total amount of code.