Get rid of REB_0_LOOKBACK, instead checking VALUE_FLAG_ENFIXED
Previously, in order to give a value a property that wouldn't be
copied when the value was copied, it could not live in the value cell
since cells were bit-copied fully.  Now that Move_Value can be
selective, what were formerly flags on keys of contexts can become
flags on the cells themselves.  This radically simplifies code for
enfixedness and protectedness.