Reuse frames in operations like ANY, ALL, CASE vs. spawn child ones
The general concept behind an operation like ANY or ALL or CASE to be
able to make a frame and reuse it was to save on frame pushing/popping
or construction when there were several successive operations.

However, an oversight led to the Do_Next_In_Frame operation always
spawning a child frame, even when an iterative frame was not holding
any important state.  This splits the operations into variants with
Do_Next_In_Subframe creating a child frame, while Do_Next_In_Frame
does not.  It also adopts the _Throws convention.

This saves considerably on frame construction and tear-down during
operations like REDUCE.  It is currently the case that the frame's
eval_type is reset to REB_0 between evaluations, this makes it safe
to keep a frame pushed on the stack even when it has no current
evaluative properties.