Reduce redundant calls to Do_Signals_Throws()
At some point a change to Do_Signals_Throws() made it get called on
every cycle.  This restores the "call only on every Eval_Dose"
behavior.

Note: The reason that the routine is called every N steps and not
just whenever a signal is set seems to have to do with implementing
the "CPU quota" feature, limiting the number of cycles the system is
allowed to run.