Do not initialize a binary with a Unicode string
MAKE_OS_BUFFER returns a unicode string, and thus it violates the
invariant "SER_WIDE(VAL_SER(v)) == 1" for binary!

Because the StdIO reading on Windows is already doing conversion from
UCS2 to UTF8, just treat it as a UTF8-encoded string in the form of
a binary.

Also added an assert in Init_Binary to check the invariant.

This fixes https://github.com/metaeducation/ren-c/issues/412