Simpler grand-unifying theory of ELSE (and THEN)
An earlier experiment introduced ELSE as an infix operator, which was
able to produce a function that could branch one of two ways.  Various
control constructs would then pass through whether a branch "succeeded"
or not, to achieve what looked like ELSE.

This particular formulation was based on the limits of infix only
being able to look one unit of code backwards, what is known as the
"tight" convention.  Hence:

    if condition [true-branch] else [false-branch]

...could only be interpreted as:

    if condition ([true-branch] else [false-branch])

While the mechanism worked, it required the cost of generating a
function each time such a branch was called...which was slow and taxed
the GC.  It also complicated the internals of the code, raising a
number of questions inside the conditional constructs as to exactly
when they should signal a call to a failing branch.  This
created an uneasy feeling, especially since IF/UNLESS/EITHER are
supposed to be simple constructs.

This revised concept uses *non-tight* evaluation, and a twist to the
rules of control constructs such that they return void if-and-only-if
no branches ran.  Hence the new interpretation of the precedence of
an ELSE expression looks more like:

    (if condition [true-branch]) else [false-branch]

As a much simpler convention, this makes ELSE work with any function
that signals its "failure" by means of returning void.  A complementary
function called THEN is added which runs only if the left hand side
is not void.

Because some code actually requires knowing or passing through a void
evaluation result on a branch, the "/?" switch is replaced with "/opt".

    if false [] => void

    if true [] => blank!
    if/opt true [] => void

IF* is the proposed shorthand for this possibly commonly-needed
operation, and as a naming scheme it is applied to the other control
constructs like SWITCH* and CASE*.  Note that it should only be
necessary to use these variations in some cases when the result of the
conditional expression is being used (by something other than ELSE).