Factor out CELL_MASK_RESET and CELL_MASK_COPY
Avoid redundancy in specifying what flags to copy or reset on a cell.
This was initially done to try and make management of
VALUE_FLAG_UNEVALUATED more automatic, and it does help in a couple of
cases (Move_Value and Derelativize now always produce a cell without
the bit set).  But it doesn't help that issue all that much.  Still
good to put the decision in one place.

Also adds macros for ANY_VAL_FLAGS() and ALL_VAL_FLAGS() to test
multiple flags at once.

Gets rid of a duplicate trashing done after allocating values from
the tail of an array (already trash).