Rename END_CELL to just END, as it's not a cell
The canon end marker is actually no longer a full cell in size, it's
just a REBNOD header.  Hence END_CELL is a deliberately misleading
name, since it doesn't have NODE_FLAG_CELL set.

This corrects an assertion from the previous commit to acknowledge
the situation when the end node is being used to prevent writes to
f->out, and also changes the name to just END.  This may have a
greater chance of collision with definitions in libraries, but it is
a macro so it could be #undef'd out of the way.