First wave of simplifying ENFIX mechanics
An attempted optimization of SET-WORD! and SET-PATH! dodged creating
new frames by instead pushing the items to the data stack.  Then, at
various points, these pending sets would be handled.  (The mechanic was
initially part of an attempt to make SET-WORD!s actually "transparent"
to the evaluator.)

The needs of enfixing have grown more complex, with a desire to have
all parameter conventions (normal, #tight, :hard-quote, 'soft-quote)
honored on the left hand argument.  Doing that requires a reworking
of the basic mechanisms for lookahead, and having this optimization
interferes with writing that code.  This rips it out for the moment,
to allow work on honing the basic features first.