Fix log parsing in tests
When a test is rerun, it failed with the error "log file parsing
problem".  The problem is that an END is passed to ANY loop in PARSE
after it finds "system/version" or a crash, and because it doesn't match
any of the valid rules, the last one:
        (do make error! "log file parsing problem")
was matched, and an error was raised

Moving the error checking after ANY fixed the issue, because the END
will break out of ANY, and match the last rule.