Remove refetch from left-quotes of SET-WORD! and SET-PATH!
When lookback quoting was first allowed on SET-WORD! and SET-PATH!,
there was some concern that authors would make the enfixed operations
taking these in their left hand side not return the same value as
what they set a variable to.  This could give rise to a situation
like `x: y: op z` where op would manipulate y, but then return a
different value for the expression...resulting in a situation where
x != y.

Trying to protect against this had technical problems, but also does
not seem very sensible when something like `quote x: 10` doesn't set
x to 10 either.  So this removes the code and puts the responsibility
on functions like DEFAULT to be sensible about returning the right
result.