Update actions.r
Limiting __write__ to not allow object! makes sense if you're doing the open, write, close single use paradigm but reduces flexibility for working with schemes.  This is to allow an object! as data, and the role of the write actor will be to make it into the form for writing to the destination.