Simplifying rewrite of REWORD, TO and THRU of BLANK! are no-ops
The REWORD mezzanine function was notoriously brittle and difficult to
understand.  @BrianH wrote a commnent saying "It's big, it's complex,
but it works"

https://github.com/rebol/rebol/blob/25033f897b2bd466068d7663563cd3ff64740b94/src/mezz/mezz-series.r#L153

In order to fix a bug in it, it seemed better to just turn it into a
much more normal Rebol function.  This simplifying change avoids many
of the problems of the original--mostly by making the match rule for
keywords remember what the keywords were, instead of having to
reverse engineer them from capturing the content out of the source
series being processed and trying to figure it back out.

Also making things simpler is the inclusion of TO or THRU of a BLANK!
value being a no-op in parse.